refactored markup removal to separate classes; this will make it easier to add html replacers (future development)
