+++ source/MerriamWebster.NET/MerriamWebster.NET/MerriamWebster.NET.csproj
@@ -13,9 +13,9 @@
     <RepositoryType>Public</RepositoryType>
     <PackageTags>Merriam-Webster dictionary thesaurus</PackageTags>
     <PackageReleaseNotes>See project readme for further details.</PackageReleaseNotes>
-    <Version>1.5.1</Version>
-    <AssemblyVersion>1.5.1</AssemblyVersion>
-    <FileVersion>1.5.1</FileVersion>
+    <Version>1.5.2</Version>
+    <AssemblyVersion>1.5.2</AssemblyVersion>
+    <FileVersion>1.5.2</FileVersion>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
