+++ source/MerriamWebster.NET/MerriamWebster.NET/Parsing/Markup/AutoLinkRemover.cs
@@ -1,0 +1,14 @@
+using System.Text.RegularExpressions;
+
+namespace MerriamWebster.NET.Parsing.Markup
+{
+    public class AutoLinkRemover : MarkupRemover, IMarkupRemover
+    {
+        public string RemoveMarkup(string input)
+        {
+            var regex = new Regex(@"{a_link\|([^}]*)}");
+
+            return RegexReplace(input, regex);
+        }
+    }
+}
