+++ source/MerriamWebster.NET/MerriamWebster.NET/Parsing/Markup/DirectLinkRemover.cs
@@ -1,0 +1,14 @@
+using System.Text.RegularExpressions;
+
+namespace MerriamWebster.NET.Parsing.Markup
+{
+    public class DirectLinkRemover : MarkupRemover, IMarkupRemover
+    {
+        public string RemoveMarkup(string input)
+        {
+            var regex = new Regex(@"{d_link\|([^}]*)\|[^}]*}");
+
+            return RegexReplace(input, regex);
+        }
+    }
+}
