+++ source/MerriamWebster.NET/MerriamWebster.NET/Parsing/Markup/EtymologyLinkRemover.cs
@@ -1,0 +1,14 @@
+using System.Text.RegularExpressions;
+
+namespace MerriamWebster.NET.Parsing.Markup
+{
+    public class EtymologyLinkRemover : MarkupRemover, IMarkupRemover
+    {
+        public string RemoveMarkup(string input)
+        {
+            var regex = new Regex(@"{et_link\|([^}]*)\|[^}]*}");
+
+            return RegexReplace(input, regex);
+        }
+    }
+}
