+++ source/MerriamWebster.NET/MerriamWebster.NET/Parsing/Markup/IMarkupRemover.cs
@@ -1,0 +1,7 @@
+namespace MerriamWebster.NET.Parsing.Markup
+{
+    public interface IMarkupRemover
+    {
+        string RemoveMarkup(string input);
+    }
+}
